# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: CIVORA API
  description: |
    CIVORA API - open digital infrastructure for public-interest service delivery.

    Version: 0.2.0-service-delivery

    ## Public-Interest Service Delivery Lifecycle

    Full end-to-end workflow for public service delivery:
    Request → Person registered → Case created (NEW) → OPEN →
    Eligibility assessed → Evidence collected → IN_REVIEW →
    Assessment completed → DECISION_PENDING → Decision made →
    APPROVED/REJECTED → IN_PROGRESS → Assistance provided →
    FOLLOW_UP → CLOSED

    State machine:
    NEW → OPEN → IN_REVIEW → ASSESSMENT → DECISION_PENDING →
    APPROVED → IN_PROGRESS → FOLLOW_UP → CLOSED
    REJECTED → CLOSED
  version: "0.2.0"
  contact:
    name: CIVORA
    url: https://github.com/alrazihi/civora
  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
servers:
  - url: http://localhost:8080
    description: Local development server. Override with CIVORA_API_URL in production.
security: []
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT bearer token. The token must include `organization_id` claim.
        The API validates that the token's organization matches the `{orgId}`
        path parameter on all tenant-scoped endpoints.
...
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: INVALID_INPUT
            message:
              type: string
              example: "title is required"
...
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          format: password
    RegisterRequest:
      type: object
      required: [email, name, password]
      properties:
        email:
          type: string
          format: email
        name:
          type: string
          maxLength: 200
        password:
          type: string
          minLength: 12
          pattern: "^(?=.*[a-zA-Z])(?=.*[0-9]).+$"
          description: "At least 12 characters with at least one letter and one number"
    LoginResponse:
      type: object
      properties:
        token:
          type: string
          description: JWT bearer token
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        user:
          type: object
          properties:
            id:
              type: string
              format: uuid
              pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
            organization_id:
              type: string
              format: uuid
              pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
            email:
              type: string
              format: email
            name:
              type: string
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        organization_id:
          type: string
          format: uuid
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        email:
          type: string
          format: email
        name:
          type: string
        role_id:
          type: string
          format: uuid
          pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
          nullable: true
        created_at:
          type: string
          format: date-time
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /organizations/{orgId}/auth/register:
    post:
      operationId: createOrgauthregister
      tags: [Authentication]
      summary: Register a new user
      description: Creates a new user within the specified organization. No prior authentication required.
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
      responses:
        '201':
          description: User registered (returns user data, not token)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '409':
          description: Email already registered in this organization
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /organizations/{orgId}/auth/login:
    post:
      operationId: createOrgauthlogin
      tags: [Authentication]
      summary: Authenticate user and get token
      description: Authenticates a user and returns a JWT bearer token.
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /organizations/{orgId}/users:
    get:
      operationId: getOrgusers
      tags: [Users]
      summary: List users
      description: Lists all users in the organization. Requires authentication.
      security:
        - BearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 20
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /organizations/{orgId}/users/{userId}:
    get:
      operationId: getOrgusersUser
      tags: [Users]
      summary: Get user
      description: Retrieves a specific user. Requires authentication.
      security:
        - BearerAuth: []
      parameters:
        - name: orgId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
            pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...